Make sure configuration sources do not get mixed up on config mutation
Currently, when `CONFIGURE SYSTEM` mutates a configuration variable,
_all_ configuration variables switch to `system override` status, which
is totally wrong.  This patch modifies the system so that the source of
each configuration option is tracked separately in a consistent manner.

Additionally, the current `DESCRIBE SYSTEM CONFIG` implementation is
wrong, because it describes the configuration as a whole without regard
for the source of the configuration values.  The correct way is to
consider `system override` values only.  While at it, I rewrote the
generation of the EdgeQL query that emits config description to make it
generic and to support configuration type subclassing properly.