Stop storing instances of schema objects in the schema container
Currently, the `Schema` object stores schema object dicts but _also_ the
Python object instances of the corresponding schema objects and schema
object collections.  This is detrimental for schema serialization as it
adds overhead when unpickling or introspecting the schema.  Fix this by
constructing the necessary objects on the fly instead.  The extra
runtime overhead is mitigated by a reasonably sized LRU cache.

This also includes some schema field getter optimizations that balance
the runtime performance hit from the above so the effect of this patch
on compilation performance is neutral.