Stricter typing for schema.get (#1900)
Get rid of the basically unconstrainted output type variable in the
type of schema.get(). Explicitly return so.Object when no `type`
argument is passed.

My motivation for this is that I had run into a few cases where the
output type was being inferred as `None` and so code wasn't getting
typechecked. And when I spot-checked some other call sites, the first
two I picked had the same problem. When I fixed the typing, though,
this didn't catch /that/ many places that had been missing
type-checking. I had to fix a handful of errors, and running with
--warn-unreachable now reports 144 errors instead of 150.

So maybe this isn't worth landing? It does give better type checking,
just not as impactfully as I had expected.