Parallelize dump tests
The current implementation of dump compatibiliy tests checks them
sequentially one by one.  Since each check is very slow, this test is,
by far, the slowest in the test suite.  Fix this by generating a test
method for each dump file, which makes it possible to run them in
parallel.