Improve handling of test isolation and parallelism
The current application of `ISOLATED_METHODS` and `SERIALIZED`
attributes of test suites is somewhat arbitrary and often misplaced
right now.

First of all, rename `SERIALIZED` to `PARALLELISM_GRANULARITY` and make
it an enum: 'default', 'database' and 'system' to signify the scope of
test side effects.  While at it, rename `ISOLATED_METHODS` to
`TRANSACTION_ISOLATION` to clarify which isolation we are talking about.

Finally, remove misplaced `TRANSACTION_ISOLATION = False` from several
test suites which used it only to avoid having to deal with error
recovery in transactions.  Use `assertRaisesRegexTx` properly instead.