Make migration constraint renames actually generate renames
Previously we would drop the constraint and create a new one,
which is bad because it requires checking the constraint for
the whole table.

Fixing this required:
 * Making parameter comparisions compare based on just the
   real param name, not the fully encoding one that includes
   the constraint/function name.
 * Having concrete constraint comparisons take into account
   in-flight renames of base constraints when comparing names.

(Should we try to move to a world where all these sorts of derived
objects don't have global derived names? They cause a lot of
grief. But probably the change would also.)

Work on #1772