Use a table instead of functions for system data
It turns out that calling an immutable constant-returning function is a
lot slower than selecting from a table if the size of the returned
datum is considerable, an order of magnitude slower, in fact.  So,
switch system data storage to use a regular table, this shaves off a
few milliseconds from the cold compiler startup time.