Refactor EdgeQL tokenization to always include the source string
When the switch to the Rusty EdgeQL lexer was made, we lost the ability
to track the original query source text, which gets in the way of
server-side error reporting.  This refactor introduces `edgeql.Source`
as the canonical way of representing an EdgeQL query source, including
the normalized version, and switches everything to use it.