Use original token stream when generating migration names
Currently, the migration name is derived from a list of tokens obtained
by deparsing DDL AST.  This is problematic, because the migration source
might not be "canonical" from the standpoint of EdgeQL codegen and so
either the client tools will have a very hard time "normalizing" the
migration source.  Avoid this by always using the original token stream,
if available.