Turn schema names into tuples
Currently, `schema.Name` is a weird subclass of `str` which adds the
`module` and the `name` attributes to it to access the name components.
There are problems with this approach: it allows mixing parsed and
unparsed names freely, since all names are strings, and also wastes
space since the name's components are stored alongside the full,
doublecolon-separated, name representation.  This is evident on the size
of the `dump01` schema dump, which is ~25% smaller as a result of this
change.