Reimplement OrderedSet in terms of regular dict
`OrderedSet` currently uses `collections.OrderedDict` as the underlying
data structure.  Since builtin dicts are now ordered, switch to using
that instead, which should be more efficient.  The removes the ability
to inject an item into the beginning of the set, but that was hardly
used.

This also switches `delta.Command` to use regular lists for subcommands,
since there no reason whatsoever to use an `OrderedSet` there.