Always return ObjectCommand from as_{create|alter|delete}_delta
The `as_create_delta` method and friends are currently declared to
declare the base `Command` instead of `ObjectCommand`, and only for the
benefit of `ReferencedInheritingObject` wanting to `DROP OWNED` before
dropping itself and returning a `CommandGroup` instead of a proper
`DeleteObject`.  There's no real reason not to nest `DROP OWNED` under
`DeleteObject` itself, so do that and annotate the delta methods
properly.