Fix migrations modifying collection aliases
Some Array fields in ArrayExprAlias weren't being serialized because
the introspection query was looking at its parent
CollectionExprAlias. Similar for tuples. Skip that.

The obvious tests for this didn't catch it because if the alias was
created during the same transaction, the reflection didn't come into
play.

Also, hide the CollectionExprAliases from reflection.