Fix functions returning a set of tuples
A tuple type returned from a function would be declared as a named
composite type in Postgres and it is an error to specify a column
definition list for such type, even if it matches the definition.
There are checks for this elsewhere, but set-returning function didn't
handle this properly.

Fixes: #2010