Refactor remote Postgres support to be more generic and robust
The current support for managed Postgres instances via `--postgres-dsn`
is limited to RDS and is rather clunky.  In order to support more
managed hosts and with less specific code, implement a notion of
"backend parameters", which include things like the cluster superuser
role and a capability bitmask that informs EdgeDB of the availability
(or lack of) certain features.

With this, EdgeDB can be bootstrapped on a DO instance and I plan to
test more providers soon.

Issue: #2006