Rework how rebasing is handled in DDL and migrations (#2018)
Currently, rebases are performed as the last operation in an ALTER,
since they are generated after everything else. (Though prior to

This changes rebases so that they can be executed at any point in a
sequence of commands inside an ALTER, which allows us to do things
like "SET OWNED on a field before dropping a base" or "add a base and
then DROP OWNED on a property declared in it", depending on how we
structure the command.

We implement this by recording the base information in AlterInherit as
we go, and then merging sequences of AlterInherit commands into Rebase
commands after the fact. All that logic moves into inheriting, which
lets us drop a nested cyclic import.

I added a group of schema tests that touch various things related to
the issue at play here, though only one of them actually failed before
this patch.

Progress on #1987: Fixes:
 * test_schema_migrations_drop_parent_03 (... which it also adds)
 * test_schema_migrations_equivalence_30

And test_constraints_exclusive_migration is getting closer...