Fix dropping inherited properties with constraints
The main bug is that inherited constraints would end up using the
subjectexpr of the abstract constraint in computing their name quals,
if the concrete parent had no subjectexpr. This resulted in it having
an incorrect name, causing it to be missed in places.

Additionally, the DDL text we generated contained subcommands nested
in a DropProperty. Prevent that, since it isn't valid in the concrete
syntax.

I think that changing how constraint names are computed means I had to
bump the version?

Progress on #1987. Fixes test_schema_migrations_equivalence_31.