Fix the inconsistent use of the DROP keyword in DDL
Currently, we use `DROP` in DDL somewhat inconsistently and confusingly to:

1. Revert to the default or inherited value  (`DROP default`).
2. In specific cases, like `DROP REQUIRED`, to set the field to a falsy value.
3. To actually delete a schema object, such as `DROP PROPERTY`.

To fix this we:

1. Introduce the `RESET <field>` syntax instead of `DROP <field>` to revert
   the field to the default or inherited value.
2. Replace uses like `DROP REQURIED` with appropriate `SET` syntax using a
   complement, specifically:

   - `DROP ABSTRACT` replaced with `SET NOT ABSTRACT`
   - `DROP FINAL` replaced with `SET NOT FINAL`
   - `DROP DELEGATED` replaced with `SET NOT DELEGATED`

The remaining use of `DROP` denoting an object deletion is thus consistent
and unambiguous.

This commit also replaces the `AlterSpecialObjectProperty` command with
a toggle on the regular `AlterObjectProperty`, which reduces the amount
of branching and will be further improved upon in upcoming patches.

The `test_edgeql_migration_eq_14b` test is xfailed temporarily due to a
DDL decompilation artifact.

Fixes: #2031