Untangle the handling of cardinality in the schema
There is currently a fair amount of brittle special-casing around the
handling of `cardinality` and `required` in pointers, specifically
around the handling of inherited and computed status, and how those
and other fields are emitted in `DESCRIBE`.

Now, there's nothing fundamentally special in `cardinality` or
`required` that prevents them to be handled in a generic fashion, except
that we need to keep track of whether the values were inferred or
specified explicitly.  For this we introduce the `computed_fields` array
that is substantially similar to the existing `inherited_fields`
attribute, but tracks the "computed" status instead.

This change has a visible effect: `DESCRIBE AS DDL|SDL` now does _not_
output the cardinality attributes if they weren't specified explicitly
in the original schema, meaning that `link foo -> Bar` would be
described as `link foo -> Bar`, and an explicit `single link foo -> Bar`
would be described as such.  This will result in a need to make a simple
migration for databases restored from previous versions of EdgeDB for
applications that use SDL schemas (in order to correctly reflect the
"inherited" and "computed" status of fields that was discarded
previously).

This patch also moves some hardcoded `DESCRIBE` bits into generic `Field`
attributes.