Fix accessing links on objects that come from functions and the like
semi_join would fail in some cases when the object came from somewhere
that didn't provide a source (like a function call or an array
indexing).

Also, drop the initial_value on array_agg, which isn't needed and
causes trouble when it can't be type inferred (such as in the bug
report for the main issue).

Closes #1887.