Remove duplication of metainformation in delta command class decls (#2058)
Delta command implementations are currently associated with their
subject schema classes via the `schema_metaclass` class argument, which
is used by `ObjectCommandMeta` to populate the registry.  The class
argument is now fully redundant, because `ObjectCommand` is a `Generic`
and we can extract the schema class from its arguments instead.

This change has a benefit of increased validation of declarations,
making sure there aren't inconsistent annotations, and, more
importantly, makes it easier to use `__init_subclass__` for registries,
which, in turn, removes the need for a few metaclasses.