Make sure computable pointers don't appear in dump
The dump/restore implementation currently erroneously considers
computable properties and links as real and includes them in the dump.
This slipped through testing because of another bug, which causes
columns to be generated for `single` properties and links. `multi`
computables, however, are broken and it is impossible to generate a dump
of a schema containing them.

Fix this by properly excluding non-dumpable pointers.  Unfortunately,
this bug causes previously-generated dumps to contain extraneous (empty)
columns, so this patch includes a procedure to rewrite the binary `COPY`
stream to eject those columns.

A portion of this patch will be backported to a7 and perhaps earlier to
make sure those versions can produce working dumps on schemas with
`multi` computables.

Issue: #2057