Fix dump consistency check
The dump table layout consistency check is incorrect with respect to
possible elided elements.  We only need to check that: a) dump tuple does
not have _extraneous_ columns (with elided elements considered), and b)
dump tuple does not have _missing_ columns (with elided elements
disregarded).

This is a followup to #2063.