Bump the graphql-core version to 3.1.2
Mainly the changes involve some variations of the error messages.
However, a new directive is introduced in this version of graphql-core:
`specifiedBy`. This is intended to be used in GraphQL schema definitions
to annotate custom scalars, which is something that could become useful
to us in the future as it addresses the issue of client code identifying
which implementaiton a given custom scalar is actually using.

There's also a change of the signature of `parse_literal` for scalars,
but the extra parameter doesn't appear to be relevant for our use case
(handling different numeric scalars).