Improve consistency of nested delta trees
When propagating ops down the inheritance tree we sometimes neglect to
construct proper subtrees that include all referrers as parent nodes,
and some places fail to establish correct CommandContext.  Fix this by
using `init_delta_branch` where appropriate and teach it to return a
stack of contexts to be used by the caller when operating on the
returned command leaf.