Improve ordering of commands in schema diffs
For each command that represents a type referred to by a link or
property in another type, set the referring type and its descendants as
weak dependents.  This helps reduce the number of unnecessary ALTERs in
the final delta, especially ones that might result in SET TYPE commands
being generated.