Fix handling of singleton bindings in external context
Many schema code paths invoke the EdgeQL compiler with a list of
bindings that should be treated in singletons.  The current formulation
achieves that by recording the specified bindings in an enveloping scope
tree branch, but then treating the first nested branch as the actual
root.  This may cause the SQL pass to miscompile queries that reference
the singleton bindings as the source relation would be left dangling
without appropriate record in the scope tree.

Fix this by recoding singleton bindings normally in the scope tree, and
passing the singleton list to the cardinality inferer directly, as was
done at some point before and the contextual infrastructure for this has
been kept around.