Improve handling of cast expressions in SET TYPE
A lack of an explicit `USING` clause in an EdgeQL `SET TYPE` statement
does not necessarily mean that one is not necessary for the underlying
SQL operation, as EdgeDB cast setup is different from Postgres,
especially when it comes to compound types.  Thus, an explicit EdgeQL
cast is injected in cases where there is no `USING` clause, or where the
result of the `USING` clause is not _exactly_ the new type.