Make scope tree unique_ids actually be unique (#2137)
We do this by getting rid of the use of "temporary" scopes in
declare_view (an approach that I previewed in #1806).

The purpose of using the temporary scopes was I think to allow the use
of the extra_scope mechanism to place the scope tree for a WITH
binding *underneath* the scope tree of a use of it. Unfortunately this
could cause duplication of unique_ids when it got grafted into two
independent subqueries, as well as some of other issues when the
fusing didn't behave right.

Fortunately, all that extra_scope machinery wasn't really needed. It
seems to work basically just fine to stick it in the scope tree where
it appears.

This fixes a handful of bugs, but also introduces some failures:
 * test_edgeql_tree_update_05 (flipping two nodes) generates invalid
   SQL, but I think fixing weak namespace stripping should fix this.
 * test_edgeql_volatility_with_01 (now split into a new test)
   fails because we now fail to detect volatile WITH bindings
   in situations like `X := random(), Y := X`. Fixing this should
   be straightforward.
 * test_edgeql_select_tvariant_04 fails because we now fail to
   deduplicate computed link sets in some cases. I haven't dug
   into this one yet.

This is work towards fixing #1381.