Fix interaction of type renames with compound and collection types
The original approach to handling the renames of the component types
of compound and collection types was to basically side-step the issue by
deriving the internal name of the compound or collection type from the
ids of the component types.  Unfortunately, because ids are
non-deterministic, this leads to the need for a bunch of unreliable hacks
to control schema diffing to avoid spurious diffs.

This patch abandons the id-based internal naming for compound and
collection types and adapts the straightforward use of component names
instead, coupled with careful rename propagation when any of the
component types gets renamed.  This is also a good opportunity to
cleanup the `RenameObject` flow in general, and so I took it, removing a
bunch of duplicative code and making `RenameObject` behave much closer
to other `AlterObjectFragment` commands.  Most significantly,
`RenameObject._canonicalize` is no longer a custom referrer recursion,
and does a regular "descend into immediate referrers and let them
propagate" action that most other commands do.

Fixes: #2002.