Implement provisional syntax for declaring extension packages
An extension package is essentially a named and versioned container for
other database objects to be declared and tracked as a unit.  The
intended use for extension packages is to bundle optional EdgeDB
functionality.  The syntax implemented here is a prototype, and not
public yet:

    CREATE EXTENSION PACKAGE <name> VERSION '<semver>' {
        <package commands>;
    };

Extension packages resemble migrations in that they also nest a block of
commands, and so this patch does some refactoring to generalize the notion
of a `NestedQLBlock`