Make Source.getptr() fallible
In the majority of cases it is useful if `Source.getptr()` fails if the
requested link or property does not exist.  So, instead of having to
`assert` the result at call sites, move the assertion error into
`getptr()` and add `maybe_get_ptr()` to be used for cases where the
lack of requested pointer is fine.