Make sure that WITH bound DML statements execute the correct number of times
This requires compiling them when we see with binding, so the
compilation happens in the correct context. The CTE will be picked up
dml_stmts when the variable is encountered.

To make FOR work, we need to also track it in dml_stmts and reuse
iterator CTEs when necessary.

Type and pointer overlays then also need to be updated whenever a DML
containing variable is used inside other DML.