Implement syntax to specify a conversion expression in SET SINGLE
When an existing link or property is converted from `multi` to
`single`, existing objects must be updated to properly reduce the
existing multi sets to a singleton value.  This implements a `USING`
clause for `SET SINGLE` (and `RESET CARDINALITY`), which acts
similarly to the analogous clause of the `SET TYPE` and `SET REQUIRED`
operations.