Don't execute nested DML when UNLESS CONFLICT fails
We accomplish this by synthesizing a SELECT that matches all
conflicting objects based on inspection of the exclusive constraints
in the object schema.

While I'm at it, I also made it work when there are subjectexprs,
which was an outstanding bug even with ON.

This still doesn't support object constraints, which will be a little
more hairy but still should be doable; I'll tackle that in a
follow-up.

Most of the way on #1669.