Separate global objects into a separate schema layer
Global objects, such as roles and databases, are shared across all
schemas in an instance, and so an addition, removal, or alteration of a
global object should become visible in all sessions.  In order to make
the introspection and maintenance of the global object set easier,
separate it into a separate `FlatSchema` layer.  Subsequent work will
build on this to make global schema synchronization more efficient.