Prevent concurrent schema mutation
There is no mechanism currently to prove that concurrent schema
mutations are not in a fundamental conflict with each other, so prohibit
_all_ concurrent schema mutations for now.  This is implemented via a
new internal sentinel `SchemaVersion` object that is altered with a new
randomly-generated UUID on each DDL statement.  The `AlterSchemaVersion`
handler then compiles a query that verifies that the original value of
`SchemaVersion` is still actual.  This prevents a race where two
mutations were compiled at the same time, and the other mutation has
been committed successfully.

Additionally, the mutation of the `SchemaVersion` relation is locked
with `SELECT FOR UPDATE` to prevent a schema mutation in concurrent
transactions.