Make UNLESS CONFLICT's handling of nested DML work with object/computed cnstrs (#2189)
We do this by collecting all the pointers that are actually needed and
then substuting them into the definition of computed pointers/object
constraints.

Also fix a silly bug where we would generate bogus SQL DDL if a
constraint key contained a space.

Fixes #1669.

I think we could make it so that ELSE works for UNLESS CONFLICT
without needing ON, now. I think it's probably worth it to do that.