Fix propagation of from_alias both in inheritance and structure (#2203)
The `from_alias` field of `Pointer` indicates whether the pointer
originates from an alias type.  Currently, propagation of this field in
inheritance when the parent type of the alias type is altered is broken.
On top of that, `from_alias` should be transitive to link properties and
it isn't.  Finally, `_propagate_ref_field_alter_in_inheritance` should
actually make the propagated field value in descendants as `inherited`.