Make the system PG connection more resilient to disconnects
The system PG connection is used for variety of things: (re)introspecting
the global state, receiving global notifications etc. It has to be
always there, always listening. With this commit we implement
a check in pgcon.connection_lost() -- if there's a server assigned
to the connection, it would notify it, so that the server can
open and setup a new system connection.