Prohibit DML in function bodies (for now)
Data modification in functions has a number of problems:

- Effects of DML inside a function is invisible to the calling query,
  which is confusing.
- The expression evaluation semantics assumes that functions have no
  side effects and so the number of invocations (0 or >1) is not
  defined.

All these problems are present in PostgreSQL as well.

To avoid surprising behavior or potential future compatibility breaks,
prohibit DML in functions altogether until a satisfactory solution to
mutating functions is implemented.

To preserve compatibility with existing databases that have mutating
functions, introduce the `allow_dml_in_functions` configuration parameter,
which enables historical behavior of allowing DML in functions and is
set automatically when a dump file from a pre-beta instance is restored.

Closes: #1661