Do not recover conn. state from the DB if COMMIT fails in parse/execute
There are many subtle differences in how parse/execute and simple query
are implemented. From the onset we tried to implement our simple query
protocol to mimic postgres' version. That caused a lot of complexity in
how we track the state, compile the queries, etc. Some of those
differences dictated the complicated design we have w.r.t. state
tracking.

The good news is that we're about to prohibit transaction control
commands in simple query, which will simplify most of the server code.
This commit makes a first step towards that: there's no need to recoved
the state from the DB if a COMMIT failed -- the state tracking in dbview
is sufficient to recover it.