Catch more errors in constraints and report them better
Disallow using EXISTS and other SET OF operations on MULTI pointers
(which was previously accepted but didn't work).

Produce a better error than "ISE: unexpectedly long path in simple
expr" for constraints that try to traverse links.

Turn the "multi-table constraints not supported" ISE into a real error
with a source context.