Handle Terminate messages appropriately
The `Terminate` protocol message (`b'X'`) is the "negotiated" connection
termination, when the client asks the server to disconnect and waits for
the connection to drop.  This ensures that both the client and the
server are in agreement with respect to whether a connection between
them still exists, which is important for cases when a command such as
`DROP DATABASE` immediately follows the client disconnect.

Unfortunately, this doesn't fully solve all `DROP DATABASE` races in the
test suite, because edgedb-python aborts connections on CancelledError
without waiting.  Until (and if) that is fixed one way or another, make
the `DROP DATABASE` retry loop unconditional.

Fixes: #2246