Discard Postgres connections on cancellation
When we issue `pg_cancel_backend`, we must discard the Postgres
connection, because we cannot be completely sure about its state.
Postgres cancellation is based on Unix signals and is addressed to the whole
worker process and not a concrete operation.  The result is that we might be
racing with the currently running query and if that completes before the
cancellation signal reaches the backend, we'll be setting a trap for the
_next_ query that is unlucky enough to pick up this Postgres backend
from the connection pool.