Fix a bunch of invalid serializations involving shapes and collections (#2274)
A common theme in these issues is an object being contained inside a
collection such that the shape can't be produced near the top-level
but instead must live inside the collection.

 * When injecting default props into a non-view object set (which
   feels sketchy, but we do it sometimes), make sure to only do it
   *once* if the set is referenced multiple times.
 * Make sure things work when an object is output with a shape and
   without a specified one in the same query, and type name injection
   is on.
 * Make sure things work when an object apears in a tuple and is
   is also referenced with a shape through tuple projection.
 * Fix array_agg and similar operating on shapes; something can
   be used as a value and serialized, so it isn't safe to
   put a serialized value as the value aspect. We need to
   compile *both*. Fixes #1818.

An earlier iteration of this PR broke test_edgeql_calls_35, because of
limitations in how we generate tuple casts. I added an xfailed test
for an already existing related failure to that bug.