Fix tuples in multi props and IN expressions
When inserting tuples into multi properties, we need to output
them as values first.

Expressions like `(1, 2) IN Foo.multi_prop` were broken because if a
tuple literal appears on the LHS of an `= ANY (` operation, pgsql
wants the subquery to contain one column per tuple field, which it won't
if we're getting the tuple straight from a table.

Instead of worrying about making the two sides agree, I just sidestepped
the issue by changing the compilation approach to compile
`x IN y` as approximately `EXISTS (SELECT ... WHERE x_output = y_output)`.

(Do we think this will cause any performance regressions? If it will
be a problem, I could instead worry about making the two sides agree.)

Also while looking at this I noticed that the optimized special-case
for `IN array_unpack(` was broken if the argument had multi
cardinality, so I fixed that also.

Fixes #2255.