Make ordering understand that DROP OWNED will delete refs
Doing this without introducing cycles in some situations required
splitting alterowned into separate setowned and dropowned actions,
from the perspective or ordering.

Also, fix the case where an annotation gets DROP OWNED and then it and
its abstract annotation is deleted. This worked in test_schema but hit
an issue in writer.

Work on #2296.