Retry racing parallel tests with intervals
Refs #2282, running the test suite on a Postgres without super user,
`CREATE ROLE` may raise a serialization error if another test is racing
for the same lock in parallel. We need some waiting time here to make
sure the retry has a higher success rate. The backoff interval algorithm
is similar to the one in edgedb-python, but waits for longer time.