Update docs
Add an explanation of how to create an exclusive constraint involving
more than one property. Give more details on the expressions that are
valid for constraints.

Clean up the reference to `START MIGRATION` in favor of deferring to the
migration tools in the SDL section.

Update the SDL/DDL documentation to emphasize SDL by moving the main
explanations there, and referring to them from DDL sections as opposed
to the other way around.

Fixes #2280
Fixes #2230