Add a generic min/max aggregate fallback
The aggregates min and max cannot handle all of our scalar types due to
how they're implemented in Postgres. So we will use specialized min/max
aggregates for the types Postgres explicitly supports and rely on a generic
catch-all implmentation for all the rest.

We introduce a `fallback` flag for function declarations. The purpose of
the flag is to designate a particular implementation as the fallback for
a polymorphic function. There's only one fallback that can be
designated. This fallback function is exempt from the restriction of
polymorphic functions using the same underlying SQL function (so that
Postgres can actually resolve the polymorphic calls). It means that this
fallback implementation will in fact be used in all cases where we
statically cannot resolve the call to a more specific version of the
polymorphic function (such as in bodies of other polymorphic functions).

Currently the `fallback` flag is not intended to be set by the users of
EdgeDB, only by the system initialization scripts that define the
standard library.

Fixes #1920

Co-authored-by: Elvis Pranskevichus <elvis@edgedb.com>