Fix handling of collection types of non-builtin scalars in dumps (#2364)
PostgreSQL encodes internal OID of element types in array and tuple
binary data.  Since only the builtin pg_catalog types have stable OIDs,
the data of arrays and tuples of all other types must be patched with
correct OID values.  We already do this in a limited form for input data
of array arguments, but the dump angle was completely overlooked.

Fixes: #2349.