Fix a staeval cast bug that produces bizarre results
const_to_python casting would directly apply the cast to the "python
value", which is wrong because tuples and None have special
meaning. Since operator eval uses a code path that uses
const_to_python on casts, this would result in concatenation of a cast
of a constant set to produce bizarre results.