Update the docs, tests, and other files to use canonical volatility.
Volatility enum follows the standard recommendation of having
"TitleCase" formatting, but UPPERCASE was used in many definitions for
backwards compatibility. However, this exception will eventually be
dropped and the proper canonical 'Immutable', 'Stable', 'Volatile"
values will have to be used. So in order to promote correct usage, the
documentation, built-in library and tests are updated to use the
standard format for enums, so that users are exposed to the correct
example as opposed to the old backwards compatibility format.