Fix creation/alteration of scalar constraints used in containers (#2407)
Migrations that created a scalar type and a property that used it
inside a container were broken for a simple reason, which is that the
constraint creation would get priority sorted after the column
creation. Tweak things around to avoid that.

However, adding a constraint to an existing scalar type that was used
in a property inside of a container type
(like `property x -> array<positive_int>`) was broken in a more
fundamental way, because pgsql disallows doing that.

We can work around this in a laborious way:
 1. Find all of the properties whose type is a container type
    that contains this scalar. (Possibly transitively.)
 2. Change the type of all offending properties to an equivalent type
    that does not reference this scalar. This may require creating
    new types.
 3. Add the constraint.
 4. Restore the type of all offending properties. If existing data
    violates the new constraint, we will fail here. Delete any
    temporarily created types.

This is a lot more code than the importance of this particular case
really ought to justify, but we live in a fallen world.

The SET TYPEs work by generating automatic conversion code using
casts. All of this infrastructure existed already but getting it
to work for the tests I wrote required fixing a couple cast bugs
and a bug in the compilation of aggregate functions used directly
as a query top-level.

Fixes #1370.