Single block: don't create connection if more than 1 available
This design balances between connection overhead and total number of
open connections when we're managing connections to only one DB.
If possible, we always try to keep 1 spare connections - no more,
no less - in the pool for the next acquire. So we don't open too many
unused connections (will be GC-ed anyway), or hold the next acquire
waiting to open new connections.

Also added base test in the simulation testing.