Add support for EdgeDB multitenancy on a single Postgres cluster
It is currently not possible to host multiple instances of EdgeDB
on the same Postgres cluster.  This complicates upgrades, since it is
not always convenient to spin up a parallel Postgres cluster just to
upgrade EdgeDB, especially if the upgrade needs to happen automatically.

Solve this by introducing a concept of EdgeDB *tenant*.  Each tenant is
given a unique identifier, which is then used to qualify all globally
visible Postgres objects (databases and roles).  It is possible to
specify the tenant id explicitly via the new `--postgres-tenant-id`
server command-line parameter.  The default tenant ID is derived from
EdgeDB catalog version, so it is now technically possible to start multiple
incompatible EdgeDB versions on the same cluster.  Currently, the server
would still refuse to do so, but upcoming work will introduce a mode
where this is allowed, and that'll make upgrades of EdgeDB running on a
remote cluster a much simpler story.

Fixes: #2158